Unify the timestamp to/from time commands
