+++ src/virtual_robots/simulators/overview_simulator.ts
@@ -4,6 +4,7 @@ import { SeededRandom } from '../../shared/base/random/seeded_random'
 import { FieldDimensions } from '../../shared/field/dimensions'
 import { message } from '../../shared/proto/messages'
 import { Ivec2 } from '../../shared/proto/messages'
+import { toTimestamp } from '../../shared/time/timestamp'
 import { Simulator } from '../simulator'
 import { Message } from '../simulator'
 import State = message.behaviour.Behaviour.State
@@ -46,7 +47,7 @@ export class OverviewSimulator implements Simulator {
     const penaltyReasons = getEnumValues<PenaltyReason>(PenaltyReason)
 
     const buffer = Overview.encode({
-      timestamp: { seconds: time },
+      timestamp: toTimestamp(time),
       robotId: index + 1,
       roleName: 'Overview Simulator',
       battery: this.random.float(),
@@ -67,9 +68,9 @@ export class OverviewSimulator implements Simulator {
       gameMode: this.random.choice(modes),
       gamePhase: this.random.choice(phases),
       penaltyReason: this.random.choice(penaltyReasons),
-      lastCameraImage: { seconds: this.randomSeconds(time, -5) },
-      lastSeenBall: { seconds: this.randomSeconds(time, -30) },
-      lastSeenGoal: { seconds: this.randomSeconds(time, -30) },
+      lastCameraImage: toTimestamp(this.randomSeconds(time, -5)),
+      lastSeenBall: toTimestamp(this.randomSeconds(time, -30)),
+      lastSeenGoal: toTimestamp(this.randomSeconds(time, -30)),
       walkCommand: {
         x: Math.cos(time / 3 + index),
         y: Math.cos(time / 5 + index),
